Allow the bulbs to change at different rates
The rate array will dictate the rate of change for the corresponding
bulb in the changers array.  Since the values in the changers array
change over time, this means that the same bulb can change at different
rates depending on the position of its index in the changers array.

Example:

In one round of the loop, changers looks like [5, 12, 33, 25, 17].
    + Bulb 5 will change at a rate of 1
    + Bulb 12 will change at a rate of 2
    + Bulb 33 will change at a rate of 3
    ...

In another round, changers looks like [17, 24, 5, 42, 19]
    + Bulb 5 will now change at a rate of 3