Change all bulbs at once & flip dim when timer out
This is a big change.  Instead of choosing a few random bulbs from the
string and modifying them until their time runs out, this update forces
all the bulbs to change at once.  When the timer on one bulb runs out,
it is reset and the dim boolean is flipped for that bulb.  In addition,
rate is no longer constant.